[WIP] Add a remove function to the example
In what will be the last version of the example, we add a remove
function to the set library. That allows to demonstrate how to ude
`state` to skew the distribution of the argument given to the functions
and to explain a bit more precisely that the `state` parameter to
the `postcond` function is the model's state before the function call.