Fix 'local' OPAM repository on Cygwin
For some reason, in CI opam accepts to add a local repository using:
- relative paths (both using forward and backward slashes)
- absolute windows paths ("file://D:\...")
but then it doesn't find any package in it.

Using absolute Cygwin paths works ("file:///cygdrive/...") so this
hard-codes where CI will put the local checkout. If that path is changed
so that the hard-coded is invalid, "opam remote add" will hopefully
error out